better fe structure
