+++ weave-js/src/components/PagePanelComponents/Home/Browse3/pages/CallPage/CallOverview.tsx
@@ -6,6 +6,7 @@ import {opNiceName} from '../common/Links';
 import {StatusChip} from '../common/StatusChip';
 import {CallSchema} from '../wfReactInterface/wfDataModelHooksInterface';
 import {ExceptionAlert} from './Exceptions';
+import {OverflowMenu} from './OverviewMenu';
 
 export const Overview = styled.div`
   display: flex;
@@ -25,6 +26,12 @@ export const CallName = styled.div`
 `;
 CallName.displayName = 'S.CallName';
 
+export const OverflowBin = styled.div`
+  align-items: right;
+  margin-left: auto;
+`;
+OverflowBin.displayName = 'S.OverflowBin';
+
 export const CallOverview: React.FC<{
   call: CallSchema;
 }> = ({call}) => {
@@ -38,6 +45,9 @@ export const CallOverview: React.FC<{
         <CallName>{opName}</CallName>
         <CallId callId={call.callId} />
         <StatusChip value={statusCode} iconOnly />
+        <OverflowBin>
+          <OverflowMenu calls={[call]} />
+        </OverflowBin>
       </Overview>
       {call.rawSpan.exception && (
         <ExceptionAlert exception={call.rawSpan.exception} />
